requirements_project.txt
streamlit==1.45.0
pandas==2.2.3
requests==2.32.3
twilio==9.6.0
python-docx==1.1.2
reportlab==4.4.0
plotly==6.0.1
openai==1.77.0
python-dotenv==1.1.0
bcrypt==4.3.0
sqlalchemy==2.0.40
psycopg2-binary